# swagger.yaml
openapi: 3.0.3
info:
  title: Proyecto Final Programacion Backend CoderHouse
  description: Este es el Proyecto Final del curso de Programacion Backend de CoderHouse. 
    Se trata de una aplicacion backend de e-commerce con la posibilidad de obtener los productos disponibles, generar usuarios con sus carritos de compra y gestionarlos, generar pedidos y enviar mensajes automaticos.
  version: 1.0.0
components:
  schemas:
    Productos:
      type: object
      required:
        - nombre
        - descripcion
        - categoria
        - precio
      properties:
        _id:
          type: string
          description: identificador unico de cada producto
        codigo:
          type: string
          description: identificador de cada producto de uso interno
        nombre:
          type: string
          description: nombre del producto
        descripcion:
          type: string
          description: descripcion del producto
        categoria:
          type: string
          description: categoria del producto
        fotos:
          description: fotos del producto
          type: array
          items:
            type: string
            description: id de la imagen            
        precio:
          type: number
          description: precio del producto en pesos argentinos
        stock:
          type: number
          description: unidades en stock del producto
      example:
        nombre: "Red Dead Redemption 2"
        descripcion: "Accion en el lejano oeste"
        foto: http://urlDeLaFotoDelProducto.jpg
        categoria: "Juegos"
        precio: 2500
        stock: 84
...
    Error:
      type: object
      properties:
        message:
          type: string
  responses:
    BadRequest:
      description: Solitud invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'  
    NotFound:
      description: El recurso solicitado no fue encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
servers:
  - url: http://localhost:8080/api
    description: Server Local
  - url: https://tp-final-backend.herokuapp.com
    description: Server de Produccion
paths:
...
  /productos:
    get:
      summary: Devuelve todos los productos
      tags:
        - Productos
      responses:
        '200':
          description: Todos los productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Productos'
    post:
      summary: Agrega un producto a la base de datos
      tags:
        - Productos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Productos'
      responses:
        '200':
          description: Creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  producto:
                    type: object
                    properties:
                      codigo:
                        type: string
                      nombre:
                        type: string
                      descripcion:
                        type: string
                      categoria:
                        type: string
                      precio:
                        type: string
                      stock:
                        type: string
                example:
                  message: Se dio de alta el producto
                  producto:
                    codigo: 8c336137-e351-4783-9fda-81a19adc6923
                    nombre: Resident Evil
                    descripcion: Terror
                    categoria: walking simulator
                    precio: '2500'
                    stock: '30'
        '401':
          description: El usuario no esta logueado/autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
              example:
                message: "Usuario no autenticado"
        '403':
          description: El usuario no tiene permisos para acceder al recurso
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Forbidden'        
              example:
                message: "No tiene permisos para acceder al recurso"
  /productos/{id}:
    get:
      summary: Devuelve el producto con el id especificado
      tags:
        - Productos
      parameters:
        - name: id
          in: path
          required: true
          description: ID del producto que queremos traer.
          schema:
            type : string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Productos'
        '404':
          description: "Producto no encontrado"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
              example:
                message: "producto no encontrado"
                id: "630f717e73bd86d8c40db5e8ee"
    delete:
      summary: Elimina un producto de la base de datos
      tags:
        - Productos
      parameters:
        - name: id
          in: path
          required: true
          description: Id del producto que se quiere eliminar
          schema:
            type: string
      responses:
        '200':
          description: Producto eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                  productos:
                    type: array
                    items:
                      $ref: '#/components/schemas/Productos'
              example:
                msg: "Producto Eliminado"
                productos:
                  - _id: "630f701273bd86d8c40db5d7"
                    nombre: "Red Dead Redemption 20"
                    descripcion: "RPG de accion"
                    categoria: "accion"
                    fotos: []
                    codigo: "724005f0-1309-4335-9c6c-cc53a187d71c"
                    precio: 2500
                    stock: 74
                    createdAt: "2022-08-31T14:28:34.789Z"
                    updatedAt: "2022-09-01T23:39:35.456Z"
                  - _id: "630f71c973bd86d8c40db5f6"
                    nombre: "Red Dead Redemption"
                    descripcion: "RPG de accion"
                    categoria: "accion"
                    fotos: []
                    codigo: "4cc8552f-a098-43cc-a2b2-1d2240002c0d"
                    precio: 2500
                    stock: 54
                    createdAt: "2022-08-31T14:35:53.424Z"
                    updatedAt: "2022-08-31T23:08:46.975Z"
        '401':
          description: El usuario no esta logueado/autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
              example:
                message: "Usuario no autenticado"
        '403':
          description: El usuario no tiene permisos para acceder al recurso
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Forbidden' 
              example:
               message: "No tiene permisos para acceder al recurso"
        '404':
          description: El producto no existe
          content:
            application/json:
              schema:
                $ref: '#/components/responses/NotFound'                
    put:
      summary: Modifica un producto de la base de datos
      tags:
        - Productos
      parameters:
        - name: id
          in: path
          required: true
          description: Id del producto que se quiere modificar
          schema:
            type: string
      requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Productos'
      responses:
        '200':
          description: Producto eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                  productos:
                    type: array
                    items:
                        $ref: '#/components/schemas/Productos'
              example:
                msg: "Producto actualizado"
                productos:
                  _id: "630f701273bd86d8c40db5d7"
                  nombre: "Red Dead Redemption 20"
                  descripcion: "RPG de accion"
                  categoria: "accion"
                  fotos: []
                  codigo: "724005f0-1309-4335-9c6c-cc53a187d71c"
                  precio: 2500
                  stock: 74
                  createdAt: "2022-08-31T14:28:34.789Z"
                  updatedAt: "2022-09-01T23:39:35.456Z"

        '401':
          description: El usuario no esta logueado/autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
              example:
                message: "Usuario no autenticado"
        '403':
          description: El usuario no tiene permisos para acceder al recurso
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Forbidden' 
              example:
                message: "No tiene permisos para acceder al recurso"
        '404':
          description: El producto no existe
          content:
            application/json:
              schema:
                $ref: '#/components/responses/NotFound'                          
...
  /carrito:
    get:
      summary: Devuelve el Carrito del usuario logueado
      tags:
        - Carritos
      parameters:
        - name: user.carrito
          in: cookie
          required: true
          description: ID del carrito del usuario logueado.
          schema:
            type : string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Carritos'
        '401':
          description: El usuario no esta logueado/autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
              example:
                message: "Usuario no autenticado"          
        '404':
          description: "Carrito no encontrado"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
              example:
                message: "Carrito no encontrado"
  /carrito/{id}:
    get:
      summary: Devuelve el Carrito del id indicado
      tags:
        - Carritos
      parameters:
        - name: id
          in: path
          required: true
          description: ID del carrito del usuario logueado.
          schema:
            type : string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Carritos'
        '401':
          description: El usuario no esta logueado/autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
              example:
                message: "Usuario no autenticado"
        '403':
          description: El usuario no tiene permisos para acceder al recurso
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Forbidden' 
              example:
                message: "No tiene permisos para acceder al recurso"
        '404':
          description: "Carrito no encontrado"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
              example:
                message: "Carrito no encontrado"
  /carrito/{id}/productos:
    get:
      summary: Devuelve el detalle de productos del Carrito del id indicado
      tags:
        - Carritos
      parameters:
        - name: id
          in: path
          required: true
          description: ID del carrito del usuario logueado.
          schema:
            type : string
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                productos:
                  - _id: "630f71d273bd86d8c40db5fc"
                    nombre: "Resident Evil Village"
                    cantidad: 15
                    precio: 1300
                  - _id: "630f701273bd86d8c40db5d7"
                    nombre: "Red Dead Redemption 2"
                    cantidad: 10
                    precio: 2500
        '401':
          description: El usuario no esta logueado/autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
              example:
                message: "Usuario no autenticado"
        '403':
          description: El usuario no tiene permisos para acceder al recurso
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Forbidden' 
              example:
                message: "No tiene permisos para acceder al recurso"
        '404':
          description: "Carrito no encontrado"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
              example:
                message: "Carrito no encontrado"
    post:
      summary: Agrega productos al Carrito indicado
      tags:
        - Carritos
      parameters:
        - name: id
          in: path
          required: true
          description: ID del carrito del usuario logueado.
          schema:
            type : string
          example: "630f6a3fc3f12bf3614f981b"
      requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  idProd:
                    type: string
                  cantidad: 
                    type: number
              example:
                id: "630f701273bd86d8c40db5d7"    
                cantidad: 3
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message: 
                    type: string
                  carrito:
                    $ref: '#/components/schemas/Carritos'
                example:
                  message: Se agrego producto al carrito
                  carrito:
                    _id: 630f6a3fc3f12bf3614f981b
                    usuario: 630f6a3fc3f12bf3614f9819
                    productos:
                      - _id: 630f701273bd86d8c40db5d7
                        nombre: Red Dead Redemption 20
                        cantidad: 21
                        precio: 2500
                    createdAt: '2022-08-31T14:03:43.349Z'
                    updatedAt: '2022-09-03T05:07:36.982Z'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
              example:
                message: 'Debe indicar un Id de Producto y cantidad para actualizar el carrito'
        '401':
          description: El usuario no esta logueado/autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
              example:
                message: "Usuario no autenticado"
        '403':
          description: El usuario no tiene permisos para acceder al recurso
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Forbidden' 
              example:
                message: "No tiene permisos para acceder al recurso"
        '404':
          description: "Carrito no encontrado"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
              example:
                message: "Carrito no encontrado"
  /carrito/productos:
    post:
      summary: Agrega productos al Carrito del usuario logueado
      tags:
        - Carritos
      parameters:
        - name: user.carrito
          in: cookie
          required: true
          description: ID del carrito del usuario logueado.
          schema:
            type : string
      requestBody:
          required: true
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: string
                  cantidad: 
                    type: number
              example:
                id: "630f701273bd86d8c40db5d7"    
                cantidad: 3
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message: 
                    type: string
                  carrito:
                    $ref: '#/components/schemas/Carritos'
                example:
                  message: Se agrego producto al carrito
                  carrito:
                    _id: 630f6a3fc3f12bf3614f981b
                    usuario: 630f6a3fc3f12bf3614f9819
                    productos:
                      - _id: 630f701273bd86d8c40db5d7
                        nombre: Red Dead Redemption 20
                        cantidad: 21
                        precio: 2500
                    createdAt: '2022-08-31T14:03:43.349Z'
                    updatedAt: '2022-09-03T05:07:36.982Z'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
              example:
                message: 'Debe indicar un Id de Producto y cantidad para actualizar el carrito'
        '401':
          description: El usuario no esta logueado/autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
              example:
                message: "Usuario no autenticado"          
        '404':
          description: "Carrito no encontrado"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
              example:
                message: "Carrito no encontrado"
    put:
      summary: decrementa la cantidad del productos al Carrito del usuario logueado
      tags:
        - Carritos
      parameters:
        - name: user.carrito
          in: cookie
          required: true
          description: ID del carrito del usuario logueado.
          schema:
            type : string
      requestBody:
          required: true
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: string
                  cantidad: 
                    type: number
              example:
                id: "630f701273bd86d8c40db5d7"    
                cantidad: 2
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message: 
                    type: string
                  carrito:
                    $ref: '#/components/schemas/Carritos'
                example:
                  message: 'Se actualizo el carrito'
                  carrito:
                    _id: 630f6a3fc3f12bf3614f981b
                    usuario: 630f6a3fc3f12bf3614f9819
                    productos:
                      - _id: 630f701273bd86d8c40db5d7
                        nombre: Red Dead Redemption 20
                        cantidad: 21
                        precio: 2500
                    createdAt: '2022-08-31T14:03:43.349Z'
                    updatedAt: '2022-09-03T05:07:36.982Z'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
              example:
                message: 'No hay tanta cantidad en el carrito'
        '401':
          description: El usuario no esta logueado/autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
              example:
                message: "Usuario no autenticado"          
        '404':
          description: "Producto no encontrado"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
              example:
                message: "Producto no se encuentra en el carrito"
                id: "630f6a3fc3f12bf3614fcd1b"
  /carrito/productos/{id_prod}:
    delete:
      summary: Elimina un producto del carrito
      tags:
        - Carritos
      parameters:
        - name: user.carrito
          in: cookie
          required: true
          description: ID del carrito
          schema:
            type : string
        - name: id_prod
          in: path
          required: true
          description: ID del producto a eliminar del carrito
          schema:
            type : string        
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message: 
                    type: string
                  carrito:
                    $ref: '#/components/schemas/Carritos'
                example:
                  message: 'Producto eliminado del carrito'
                  carrito:
                    _id: 630f6a3fc3f12bf3614f981b
                    usuario: 630f6a3fc3f12bf3614f9819
                    productos:
                      - _id: 630f701273bd86d8c40db5d7
                        nombre: Red Dead Redemption 20
                        cantidad: 21
                        precio: 2500
                    createdAt: '2022-08-31T14:03:43.349Z'
                    updatedAt: '2022-09-03T05:07:36.982Z'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/responses/BadRequest'
              example:
                message: 'Debe indicar un Id de Producto para eliminar'
        '401':
          description: El usuario no esta logueado/autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
              example:
                message: "Usuario no autenticado"          
        '404':
          description: "Carrito no encontrado"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
              example:
                message: "Carrito no encontrado"
                id: "630f6a3fc3f12bf3614fcd1b"
...